@@ -70,6 +70,8 @@
     "fetch": "^1.1.0",
     "file-loader": "^0.9.0",
     "fs-extra": "^0.30.0",
+    "highcharts": "^5.0.14",
+    "highstock-release": "^5.0.14",
     "history": "^2.0.0",
     "html-webpack-plugin": "^2.7.1",
     "immutability-helper": "^2.2.2",
@@ -87,20 +89,23 @@
     "koa-static": "^3.0.0",
     "material-ui": "^0.16.4",
     "moment": "^2.18.1",
-    "node-fetch": "^1.6.3",
+    "node-fetch": "^1.7.2",
     "node-sass": "^3.7.0",
     "normalize.css": "^4.1.1",
     "postcss-loader": "^0.9.0",
     "prop-types": "^15.5.10",
-    "react": "^15.4.1",
+    "react": "^15.6.1",
     "react-bootstrap-toggle": "^2.0.4",
     "react-color": "^2.11.7",
     "react-date-picker": "^5.3.28",
     "react-date-range": "^0.2.4",
-    "react-dom": "^15.4.1",
+    "react-dom": "^15.6.1",
     "react-file-uploader": "^0.3.3",
     "react-google-login": "^2.3.2",
     "react-google-login-component": "^0.5.2",
+    "react-jsx-highcharts": "^1.3.4",
+    "react-jsx-highstock": "^1.3.4",
+    "react-month-picker": "^1.3.0",
     "react-redux": "^4.0.0",
     "react-router": "^2.2.0",
     "react-router-redux": "^4.0.0",
