@@ -45,6 +45,8 @@ import Page_mail_template from './modules/templates/containers/addTemplate';
 import Page_AddVariables from './modules/templates/containers/addVariables';
 import Page_TeamView from './modules/team/containers/viewTeam';
 import Page_MyProfile from './modules/myProfile/containers/myProfile';
+import PageManageDashboard from 'modules/manageUsers/containers/manageDashboard';
+import Page_AttendanceReq from './modules/attendance/containers/manageAttendanceReq';
 
 import store from './store';
 // import APP from './App';
@@ -72,9 +74,11 @@ let render = (routerKey = null) => {
       <Provider store={store}>
         <Router history={appHistory} queryKey={false}>
           <Route path="/" component={APP}>
-            <IndexRoute component={Page_Login} />
+            <IndexRoute component={PageManageDashboard} />
             //this will be the default page which will opens when app starts
             <Route path="home" component={Page_Home} />
+            <Route path="PageManageDashboard" component={PageManageDashboard} />
+            <Route path="page_login" component={Page_Login} />
             <Route path="monthly_attendance" component={Page_MonthlyAttendance} />
             <Route path="manage_working_hours" component={Page_ManageWorkingHours} />
             <Route path="logout" component={Page_Logout} />
@@ -103,6 +107,7 @@ let render = (routerKey = null) => {
             <Route path="add_variables" component={Page_AddVariables} />
             <Route path="mail_templates" component={Page_mail_template} />
             <Route path="inventory_system" component={Page_InventorySystem} />
+            <Route path="attendanceReq" component={Page_AttendanceReq} />
           </Route>
         </Router>
       </Provider>
