@@ -9,6 +9,7 @@ import Menu from 'components/generic/Menu';
 import {isNotUserValid} from 'src/services/generic';
 import Header from 'components/generic/Header';
 import UsersList from 'components/generic/UsersList';
+import UsersListHeader from 'components/generic/UsersListHeader';
 import UserSalaryHistory from 'components/salary/manageSalary/UserSalaryHistory';
 import UserHoldingHistory from 'components/salary/manageSalary/UserHoldingHistory';
 import FormAddHolding from 'modules/salary/components/manageSalary/FormAddHolding';
@@ -153,15 +154,16 @@ class ManageSalary extends React.Component {
       <div>
         <Menu {...this.props} />
         <div id="content" className="app-content box-shadow-z0" role="main">
-          <Header pageTitle={'Manage Salaries' + status_message} showLoading={this.props.frontend.show_loading} />
+          <Header pageTitle={'Manage Salaries' + status_message} showLoading={this.props.frontend.show_loading} userListHeader />
+          <UsersListHeader users={this.props.usersList.users} selectedUserId={this.state.selected_user_id} onUserClick={this.onUserClick} />
           <div className="app-body" id="view">
             <div className="padding">
               <div className="row">
-                <div className="col-md-2" id="fixedScroll">
-                  <UsersList users={this.props.loggedUser.data.role == CONFIG.HR ? this.state.subList : this.props.usersList.users} selectedUserId={this.state.selected_user_id} onUserClick={this.onUserClick} {...this.props} top={10} />
+                <div className="col-md-2 col-sm-12 col-xs-12 hidden-xs" id="fixedScroll">
+                  <UsersList users={this.props.loggedUser.data.role == CONFIG.HR ? this.state.subList : this.props.usersList.users} selectedUserId={this.state.selected_user_id} onUserClick={this.onUserClick} {...this.props} />
                 </div>
-                <div className="col-md-10">
-                  <div className="box">
+                <div className="col-md-10 col-sm-12 col-xs-12">
+                  <div className="box m-t-xs">
                     <div className="p-a text-center">
                       <a href="" className="text-md m-t block">{this.state.selected_user_name}</a>
                       <p>
@@ -170,21 +172,21 @@ class ManageSalary extends React.Component {
                     </div>
                   </div>
                   <div className="row no-gutter b-t box">
-                    <div className="col-xs-3 b-r box">
+                    <div className="col-md-3 col-sm-4 col-xs-12 b-r box">
                       <div className="p-a block ">
                         <h6 className="text-center">Salary Revision</h6>
                         <hr />
                         <UserSalaryHistory data={this.state.salary_history} message={this.state.msg} viewSalarySummary={this.viewSalarySummary} callDeleteUserSalary={this.callDeleteUserSalary} />
                       </div>
                     </div>
-                    <div className="col-xs-6 b-r box">
+                    <div className="col-md-6 col-sm-8 col-xs-12 b-r box">
                       <div className="p-a block">
                         <h6 className="text-center">Add New</h6>
                         <hr />
                         <FormAddSalary {...this.props} userid={this.state.selected_user_id} callAddUserSalary={this.callAddUserSalary} user_latest_salary_details={this.state.user_latest_salary_details} />
                       </div>
                     </div>
-                    <div className="col-xs-3 b-r box">
+                    <div className="col-md-3 col-sm-8 col-xs-12 b-r box">
                       <div className="p-a block ">
                         <h6 className="text-center">Holding Revision</h6>
                         <hr />
