@@ -25,8 +25,7 @@ class ViewTeam extends React.Component {
     this.openPage = this.openPage.bind(this);
   }
   componentWillMount () {
-    this.props.requestTeamList();
-    this.props.requestGetTeam('');
+    this.props.requestUserDayAttendance();
   }
   componentWillReceiveProps (props) {
     window.scrollTo(0, 0);
