@@ -24,6 +24,15 @@ export const REQUEST_USER_ATTENDANCE = 'REQUEST_USER_ATTENDANCE';
 export const SUCCESS_USER_ATTENDANCE = 'ACTION_SUCCESS_USER_ATTENDANCE';
 export const ERROR_USER_ATTENDANCE = 'ACTION_ERROR_USER_ATTENDANCE';
 
+// attendence REQUEST
+export const REQUEST_USER_ATTENDANCE_REQUEST = 'REQUEST_USER_ATTENDANCE_REQUEST';
+export const SUCCESS_USER_ATTENDANCE_REQUEST = 'ACTION_SUCCESS_USER_ATTENDANCE_REQUEST';
+export const ERROR_USER_ATTENDANCE_REQUEST = 'ACTION_ERROR_USER_ATTENDANCE_REQUEST';
+// attendence status
+export const REQUEST_USER_ATTENDANCE_STATUS = 'REQUEST_USER_ATTENDANCE_STATUS';
+export const SUCCESS_USER_ATTENDANCE_STATUS = 'ACTION_SUCCESS_USER_ATTENDANCE_STATUS';
+export const ERROR_USER_ATTENDANCE_STATUS = 'ACTION_ERROR_USER_ATTENDANCE_STATUS';
+
 // get user day summary
 export const REQUEST_USER_DAY_SUMMARY = 'REQUEST_USER_DAY_SUMMARY';
 export const SUCCESS_USER_DAY_SUMMARY = 'ACTION_SUCCESS_USER_DAY_SUMMARY';
@@ -34,6 +43,10 @@ export const REQUEST_UPDATE_USER_DAY_SUMMARY = 'REQUEST_UPDATE_USER_DAY_SUMMARY'
 export const SUCCESS_UPDATE_USER_DAY_SUMMARY = 'ACTION_SUCCESS_UPDATE_USER_DAY_SUMMARY';
 export const ERROR_UPDATE_USER_DAY_SUMMARY = 'ACTION_ERROR_UPDATE_USER_DAY_SUMMARY';
 
+// user update day summary --
+export const REQUEST_UPDATE_EMP_DAY_SUMMARY = 'REQUEST_UPDATE_EMP_DAY_SUMMARY';
+export const SUCCESS_UPDATE_EMP_DAY_SUMMARY = 'ACTION_SUCCESS_UPDATE_EMP_DAY_SUMMARY';
+export const ERROR_UPDATE_EMP_DAY_SUMMARY = 'ACTION_ERROR_UPDATE_EMP_DAY_SUMMARY';
 // USERS LIST
 export const REQUEST_USERSLIST = 'ACTION_EMPTY_USERSLIST';
 export const SUCCESS_USERSLIST = 'ACTION_SUCCESS_USERSLIST';
@@ -67,13 +80,38 @@ export const REQUEST_GET_TEAM = 'REQUEST_GET_TEAM';
 export const SUCCESS_GET_TEAM = 'SUCCESS_GET_TEAM';
 export const ERROR_GET_TEAM = 'ERROR_GET_TEAM';
 
+export const REQUEST_TEAM_STATS = 'REQUEST_TEAM_STATS';
+export const SUCCESS_TEAM_STATS = 'SUCCESS_TEAM_STATS';
+export const ERROR_TEAM_STATS = 'ERROR_TEAM_STATS';
+
+export const REQUEST_EMP_LIFE_CYCLE = 'REQUEST_EMP_LIFE_CYCLE';
+export const SUCCESS_EMP_LIFE_CYCLE = 'SUCCESS_EMP_LIFE_CYCLE';
+export const ERROR_EMP_LIFE_CYCLE = 'ERROR_EMP_LIFE_CYCLE';
+
+export const REQUEST_EMP_HOURS = 'REQUEST_EMP_HOURS';
+export const SUCCESS_EMP_HOURS = 'SUCCESS_EMP_HOURS';
+export const ERROR_EMP_HOURS = 'ERROR_EMP_HOURS';
+
+export const REQUEST_EMP_MONTHLY_HOURS = 'REQUEST_EMP_MONTHLY_HOURS';
+export const SUCCESS_EMP_MONTHLY_HOURS = 'SUCCESS_EMP_MONTHLY_HOURS';
+export const ERROR_EMP_MONTHLY_HOURS = 'ERROR_EMP_MONTHLY_HOURS';
+
+export const REQUEST_EMP_PERFORMANCE = 'REQUEST_EMP_PERFORMANCE';
+export const SUCCESS_EMP_PERFORMANCE = 'SUCCESS_EMP_PERFORMANCE';
+export const ERROR_EMP_PERFORMANCE = 'ERROR_EMP_PERFORMANCE';
+
+// USER List
+export const REQUEST_USER_LIST = 'REQUEST_USER_LIST';
+export const SUCCESS_USER_LIST = 'SUCCESS_USER_LIST';
+export const ERROR_USER_LIST = 'ERROR_USER_LIST';
+
 //* **********************************************************************************
 // +++""ADMIN""+++
 
 export const ACTION_SUCCESS_DISABLED_USERSLIST = 'ACTION_SUCCESS_DISABLED_USERSLIST';
 export const ACTION_ERROR_DISABLED_USERSLIST = 'ACTION_ERROR_DISABLED_USERSLIST';
 
-// 'INVENTORY'
+// 'INVENTORY
 export const ACTION_SUCCESS_ADD_NEW_MACHINE = 'ACTION_SUCCESS_ADD_NEW_MACHINE';
 export const ACTION_ERROR_ADD_NEW_MACHINE = 'ACTION_ERROR_ADD_NEW_MACHINE';
 
