@@ -25,7 +25,10 @@ import {empSalaryList} from 'appRedux/salary/reducers/empSalaryDetails';
 import policyDocuments from 'appRedux/policyDocuments/reducers/policyDocuments';
 import {template} from 'appRedux/templates/reducers/template';
 import teamList from 'appRedux/team/reducers/teamList';
+import teamStats from 'appRedux/manageUsers/reducers/dashboard';
 import {manageDevice} from 'appRedux/inventory/reducers/inventory';
+import attendanceReq from 'appRedux/attendance/reducers/attendanceReq';
+import empDaySummary from 'appRedux/attendance/reducers/empDaySummary';
 
 export default combineReducers({
   frontend,
@@ -53,5 +56,8 @@ export default combineReducers({
   policyDocuments,
   template,
   teamList,
-  manageDevice
+  teamStats,
+  manageDevice,
+  attendanceReq,
+  empDaySummary
 });
